Move image annotator viz to charts platform
