+++ config/plugins/visualizations/annotate_image/config/annotate_image.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE visualization SYSTEM "../../visualization.dtd">
-<visualization name="Annotate image">
+<visualization name="Image annotator">
     <description>An image annotater built using PaperJS at https://github.com/paperjs/paper.js.</description>
     <data_sources>
         <data_source>
@@ -13,5 +13,5 @@
     <params>
         <param type="dataset" var_name_in_template="hda" required="true">dataset_id</param>
     </params>
-    <template>annotate_image.mako</template>
+    <entry_point entry_point_type="chart" src="script.js" css="jquery.contextMenu.css"/>
 </visualization>
