+++ config/plugins/visualizations/annotate_image/package.json
@@ -1,6 +1,7 @@
 {
-    "name": "visualization",
+    "name": "annotate_image",
     "version": "0.1.0",
+    "description": "Image annotator visualisation using paperJS",
     "keywords": [
         "galaxy",
         "visualization"
@@ -9,7 +10,12 @@
     "dependencies": {
         "babel-preset-env": "^1.6.1",
         "backbone": "^1.3.3",
-        "parcel-bundler": "^1.4.1"
+        "jquery": "^3.3.1",
+        "jquery-contextmenu": "^2.7.1",
+        "jquery.ui.position": "^1.11.4",
+        "paper": "^0.12.3",
+        "parcel-bundler": "^1.4.1",
+        "underscore": "^1.8.3"
     },
     "scripts": {
         "build": "parcel build src/script.js -d static"
