--- config/plugins/visualizations/annotate_image/static/css/create-canvas.css
@@ -1,17 +1,0 @@
-﻿body {
-}
-
-.img-container {
-    position: relative;
-}
-
-.image-canvas {
-    cursor: url(../images/red-pen.png) 14 50, auto;
-}
-
-.context-menu-item.context-menu-icon-text { background-image: url(../images/text.png); }
-.context-menu-item.context-menu-icon-blackpen { background-image: url(../images/blackpen.png); }
-.context-menu-item.context-menu-icon-redpen { background-image: url(../images/redpen.png); }
-.context-menu-item.context-menu-icon-greenpen { background-image: url(../images/greenpen.png); }
-.context-menu-item.context-menu-icon-bluepen { background-image: url(../images/bluepen.png); }
-.context-menu-item.context-menu-icon-yellowpen { background-image: url(../images/yellowpen.png); }
